Remove all the old catalog code
