+++ .github/CODEOWNERS
@@ -15,14 +15,12 @@
 /src/libs/ajax/billing/ @DataBiosphere/broad-core-services
 /src/libs/ajax/GoogleStorage.js @DataBiosphere/dsp-analysis
 /src/libs/ajax/Groups.ts @DataBiosphere/broad-core-services
-/src/libs/ajax/Catalog.ts @DataBiosphere/dsp-data-team-eng
 /src/libs/ajax/workflows-app/ @DataBiosphere/dsp-analysis
 /src/libs/ajax/workspaces/ @DataBiosphere/broad-core-services
 /src/pages/library/DataBrowser* @DataBiosphere/dsp-data-team-eng
 /src/pages/library/Datasets* @DataBiosphere/dsp-data-team-eng
 /src/pages/library/SearchAndFilterComponent* @DataBiosphere/dsp-data-team-eng
 /src/dataset-builder/ @DataBiosphere/dsp-data-team-eng
-/src/data-catalog/ @DataBiosphere/dsp-data-team-eng
 /src/pages/LandingPage* @DataBiosphere/terra-cobranding
 /src/pages/billing/ @DataBiosphere/broad-core-services
 /src/pages/groups/ @DataBiosphere/broad-core-services
@@ -38,8 +36,4 @@
 /integration-tests/tests/run-rstudio.js @DataBiosphere/dsp-analysis
 /integration-tests/tests/analysis-context-bar.js @DataBiosphere/dsp-analysis
 /integration-tests/tests/export-*-dataset-to-workspace.js @DataBiosphere/dsp-data-team-eng
-/integration-tests/tests/preview-dataset.js @DataBiosphere/dsp-data-team-eng
-/integration-tests/tests/request-access.js @DataBiosphere/dsp-data-team-eng
-/integration-tests/tests/run-catalog* @DataBiosphere/dsp-data-team-eng
-/integration-tests/utils/catalog-utils.js @DataBiosphere/dsp-data-team-eng
 /integration-tests/utils/import-tdr-snapshot.ts @DataBiosphere/broad-core-services
