+++ config/staging.json
@@ -3,7 +3,6 @@
   "bardRoot": "https://terra-bard-staging.appspot.com",
   "billingProfileManagerUrlRoot": "https://bpm.dsde-staging.broadinstitute.org",
   "calhounUrlRoot": "https://calhoun.dsde-staging.broadinstitute.org",
-  "catalogUrlRoot": "https://catalog.dsde-staging.broadinstitute.org",
   "dataRepoUrlRoot": "https://data.staging.envs-terra.bio",
   "devUrlRoot": "https://bvdp-saturn-staging.appspot.com",
   "dockstoreUrlRoot": "https://staging.dockstore.org",
