+++ integration-tests/slack/slack-notify-channels.json
@@ -42,24 +42,6 @@
       "id": "C07T2DLHD24",
       "tests": [
         {
-          "name": "preview-dataset"
-        },
-        {
-          "name": "run-catalog-filter"
-        },
-        {
-          "name": "export-tdr-dataset-to-workspace"
-        },
-        {
-          "name": "export-workspace-dataset-to-workspace"
-        },
-        {
-          "name": "import-cohort-data"
-        },
-        {
-          "name": "request-access"
-        },
-        {
           "name": "preview-drs-uri"
         }
       ]
