+++ src/dataset-builder/DatasetBuilder.ts
@@ -9,7 +9,7 @@ import { ValidatedInput } from 'src/components/input';
 import { MenuButton } from 'src/components/MenuButton';
 import { makeMenuIcon, MenuTrigger } from 'src/components/PopupTrigger';
 import { TopBar } from 'src/components/TopBar';
-import { StringInput } from 'src/data-catalog/create-dataset/CreateDatasetInputs';
+import { StringInput } from 'src/dataset-builder/custom-inputs/CreateDatasetInputs';
 import {
   addSelectableObjectToGroup,
   Cohort,
