+++ src/import-data/ImportDataOverview.ts
@@ -36,7 +36,6 @@ const getTitleForImportRequest = (importRequest: ImportRequest): string => {
     case 'tdr-snapshot-export':
       return `Import snapshot ${importRequest.snapshot.name}`;
     case 'tdr-snapshot-reference':
-    case 'catalog-dataset':
       return 'Link data to a workspace';
     default:
       return 'Import data to a workspace';
