+++ src/import-data/useImportRequest.test.ts
@@ -1,12 +1,10 @@
 import { partial } from '@terra-ui-packages/test-utils';
-import { fetchDataCatalog } from 'src/data-catalog/data-browser-utils';
 import { DataRepo, DataRepoContract, Snapshot } from 'src/libs/ajax/DataRepo';
 import { SamResources, SamResourcesContract } from 'src/libs/ajax/SamResources';
 import { asMockedFn } from 'src/testing/test-utils';
 
 import {
   BagItImportRequest,
-  CatalogDatasetImportRequest,
   EntitiesImportRequest,
   ImportRequest,
   PFBImportRequest,
@@ -31,14 +29,6 @@ jest.mock('src/libs/ajax/SamResources', (): SamResourcesExports => {
   return {
     ...jest.requireActual<SamResourcesExports>('src/libs/ajax/SamResources'),
     SamResources: jest.fn(),
-  };
-});
-
-type DataBrowserUtilsExports = typeof import('src/data-catalog/data-browser-utils');
-jest.mock('src/data-catalog/data-browser-utils', (): DataBrowserUtilsExports => {
-  return {
-    ...jest.requireActual<DataBrowserUtilsExports>('src/data-catalog/data-browser-utils'),
-    fetchDataCatalog: jest.fn(),
   };
 });
 
@@ -134,17 +124,6 @@ describe('getImportRequest', () => {
         snapshotAccessControls: [],
       } satisfies TDRSnapshotReferenceImportRequest,
     },
-    // Catalog dataset
-    {
-      queryParams: {
-        format: 'catalog',
-        catalogDatasetId: '00001111-2222-3333-aaaa-bbbbccccdddd',
-      },
-      expectedResult: {
-        type: 'catalog-dataset',
-        datasetId: '00001111-2222-3333-aaaa-bbbbccccdddd',
-      } satisfies CatalogDatasetImportRequest,
-    },
   ];
 
   const mockSnapshotDetails = (snapshotId: string) =>
@@ -162,57 +141,6 @@ describe('getImportRequest', () => {
       }),
     };
     asMockedFn(DataRepo).mockReturnValue(mockDataRepo as unknown as DataRepoContract);
-
-    asMockedFn(fetchDataCatalog).mockResolvedValue([
-      {
-        id: 'aaaabbbb-cccc-dddd-eeee-ffffgggghhhh',
-        'dct:creator': 'testowner',
-        'dct:description': 'A test snapshot',
-        'dct:identifier': '00001111-2222-3333-aaaa-bbbbccccdddd',
-        'dct:issued': '2023-10-02T11:30:00.000000Z',
-        'dct:title': 'test-snapshot-1',
-        'dcat:accessURL':
-          'https://jade.datarepo-dev.broadinstitute.org/snapshots/details/00001111-2222-3333-aaaa-bbbbccccdddd',
-        'TerraDCAT_ap:hasDataCollection': [],
-        accessLevel: 'reader',
-        storage: [],
-        counts: {},
-        samples: {},
-        contributors: [],
-      },
-      {
-        id: '11112222-3333-4444-5555-666677778888',
-        'dct:creator': 'testowner',
-        'dct:description': 'Another test snapshot',
-        'dct:identifier': 'aaaabbbb-cccc-1111-2222-333333333333',
-        'dct:issued': '2023-10-02T11:30:00.000000Z',
-        'dct:title': 'test-snapshot-2',
-        'dcat:accessURL':
-          'https://jade.datarepo-dev.broadinstitute.org/snapshots/details/aaaabbbb-cccc-1111-2222-333333333333',
-        'TerraDCAT_ap:hasDataCollection': [],
-        accessLevel: 'reader',
-        storage: [],
-        counts: {},
-        samples: {},
-        contributors: [],
-      },
-      {
-        id: 'zzzzyyyy-xxxx-1111-2222-333333333333',
-        'dct:creator': 'testowner',
-        'dct:description': 'Yet another test snapshot',
-        'dct:identifier': '99998888-7777-xxxx-yyyy-zzzzzzzzzzzz',
-        'dct:issued': '2023-10-02T11:30:00.000000Z',
-        'dct:title': 'test-snapshot-3',
-        'dcat:accessURL':
-          'https://jade.datarepo-dev.broadinstitute.org/snapshots/details/99998888-7777-xxxx-yyyy-zzzzzzzzzzzz',
-        'TerraDCAT_ap:hasDataCollection': [],
-        accessLevel: 'reader',
-        storage: [],
-        counts: {},
-        samples: {},
-        contributors: [],
-      },
-    ]);
 
     asMockedFn(SamResources).mockReturnValue(
       partial<SamResourcesContract>({
