+++ src/libs/events.ts
@@ -66,14 +66,6 @@ const eventsList = {
   cloudEnvironmentDelete: 'cloudEnvironment:delete',
   cloudEnvironmentUpdate: 'cloudEnvironment:update',
   cloudEnvironmentDetailsLoad: 'analysis:details:load',
-  catalogFilterSearch: 'catalog:filter:search',
-  catalogFilterSidebar: 'catalog:filter:sidebar',
-  catalogRequestAccess: 'catalog:requestAccess',
-  catalogToggle: 'catalog:toggle',
-  catalogViewDetails: 'catalog:view:details',
-  catalogViewPreviewData: 'catalog:view:previewData',
-  catalogWorkspaceLinkFromDetailsView: 'catalog:workspaceLink:detailsView',
-  catalogWorkspaceLinkExportFinished: 'catalog:workspaceLink:completed',
   datasetLibraryBrowseData: 'library:browseData',
   dataTableOpenUploader: 'dataTable:openUploader',
   dataTableSaveColumnSettings: 'dataTable:saveColumnSettings',
